Code enhanced controller

+++ alj-java-challenge-master/src/main/java/jp/co/axa/apidemo/controllers/EmployeeController.java
@@ -7,6 +7,7 @@ import jp.co.axa.apidemo.services.EmployeeService;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
@@ -14,6 +15,7 @@ import org.springframework.web.server.ResponseStatusException;
 import javax.validation.constraints.NotNull;
 import java.util.List;
 
+import static jp.co.axa.apidemo.validatorutils.ValidatorUtils.isNull;
 import static jp.co.axa.apidemo.validatorutils.ValidatorUtils.validateEmployee;
 
 /**
@@ -47,7 +49,11 @@ public class EmployeeController {
      */
     @GetMapping(value = {"/employees/", "/employees/{employeeId}",})
     public Employee getEmployee(@PathVariable(name = "employeeId") Long employeeId) {
-        return employeeService.getEmployee(employeeId);
+        Employee emp = employeeService.getEmployee(employeeId);
+        if (isNull(emp)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Employee not exist");
+        }
+        return emp;
     }
 
     /**
@@ -56,16 +62,22 @@ public class EmployeeController {
     @PostMapping("/employees")
     public ResponseEntity<Employee> saveEmployee(@RequestBody @NotNull EmployeeRequest employeeRequest) {
         Employee employee = new Employee();
-        try {
             /**
              * validating employeeRequest object sent from request body
              */
             validateEmployee(employeeRequest);
-
             /**
              * Mapping employeeRequest to employee Entity
              * Persistent entities should not be used as arguments of @RequestMapping
              */
+            if(employeeRequest.getId() > 0) {
+                Employee emp = employeeService.getEmployee(employeeRequest.getId());
+                if (!isNull(emp))
+                    throw new EmployeeApiException("Employee Id is already exists");
+                employee.setId(employeeRequest.getId());
+            } else if ((employeeRequest.getId() == 0) || (employeeRequest.getId() < 0)) {
+                throw new EmployeeApiException("Employee Id should be greater than 0 or non-negative");
+            }
             employee.setName(employeeRequest.getName());
             employee.setSalary(employeeRequest.getSalary());
             employee.setDepartment(employeeRequest.getDepartment());
@@ -74,23 +86,10 @@ public class EmployeeController {
              * After Successful validation and mapping save the employee
              */
             employeeService.saveEmployee(employee);
-        } catch (EmployeeApiException e) {
-            /**
-             * Catch Employee API exceptions thrown during validations
-             */
-            log.info(e.getMessage());
-            throw new EmployeeApiException(e.getMessage());
-        } catch (Exception e) {
-            /**
-             * Catch Unexpected API exceptions and throw custom API Exception
-             */
-            log.debug(e.getCause());
-            throw new EmployeeApiException("Internal error");
-        }
 
         log.info("Employee Saved Successfully");
 
-        return ResponseEntity.status(HttpStatus.CREATED).body(employee);
+        return ResponseEntity.status(HttpStatus.CREATED).contentType(MediaType.APPLICATION_JSON).body(employee);
     }
 
     /**
@@ -98,34 +97,21 @@ public class EmployeeController {
      */
     @DeleteMapping("/employees/{employeeId}")
     public ResponseEntity<String> deleteEmployee(@PathVariable(name = "employeeId") Long employeeId) {
-        try {
             /**
              * To delete employee first check that employee
              *  already exists by retrieve employee
              *  if not exists it will throw custom API exception
              */
-            employeeService.getEmployee(employeeId);
+            Employee emp = employeeService.getEmployee(employeeId);
 
             /**
              * To delete employee by employeeId
              */
-            employeeService.deleteEmployee(employeeId);
-        } catch (EmployeeApiException e) {
-            /**
-             * Catch Employee API exception if employee
-             * user want to delete not exists
-             */
-            log.info("Employee deletion failed");
-            throw new EmployeeApiException(e.getMessage());
-        } catch (Exception e) {
-            /**
-             * Catch Unexpected API exception
-             */
-            log.debug(e.getCause());
-            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Internal server error");
-        }
-        log.info("Employee Deleted Successfully");
-        return ResponseEntity.status(HttpStatus.OK).body("Deleted Successfully");
+            if (!isNull(emp)) {
+                employeeService.deleteEmployee(employeeId);
+                log.info("Employee Deleted Successfully");
+            }
+        return ResponseEntity.status(HttpStatus.OK).contentType(MediaType.APPLICATION_JSON).body("Deleted Successfully");
     }
 
     /**
@@ -134,7 +120,6 @@ public class EmployeeController {
     @PutMapping("/employees/{employeeId}")
     public ResponseEntity<String> updateEmployee(@RequestBody @NotNull EmployeeRequest employeeRequest,
                                                  @PathVariable(name = "employeeId") Long employeeId) {
-        try {
             /**
              * Check that the employee user wants to
              *  update is exis or not
@@ -157,19 +142,7 @@ public class EmployeeController {
              * Update employee
              */
             employeeService.updateEmployee(emp);
-        } catch (EmployeeApiException e) {
-            /**
-             * Catch Employee API exceptions thrown during validations
-             */
-            log.info("Update employee failed");
-            throw new EmployeeApiException(e.getMessage());
-        } catch (Exception e) {
-            /**
-             * Catch Unexpected API exception
-             */
-            log.debug(e.getCause());
-            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Internal error");
-        }
-        return ResponseEntity.status(HttpStatus.OK).body("Update Success");
+
+        return ResponseEntity.status(HttpStatus.OK).contentType(MediaType.APPLICATION_JSON).body("Update Success");
     }
 }
